Feat: Introduce router logic and create a package for all routers and put them inside
