+++ app/src/main/java/com/android/apomden/Fragments/Routers/ProfileRouterFragment.java
@@ -1,18 +1,17 @@
-package com.android.apomden.Fragments;
+package com.android.apomden.Fragments.Routers;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.RecyclerView;
 import androidx.viewpager.widget.ViewPager;
 
-import com.android.apomden.Adapters.BedRecyclerAdapter;
 import com.android.apomden.Adapters.SectionsPagerAdapter;
+import com.android.apomden.Fragments.IncomingTransferFragment;
+import com.android.apomden.Fragments.OutgoingTransferFragment;
 import com.android.apomden.R;
 import com.google.android.material.tabs.TabLayout;
 
@@ -20,22 +19,17 @@ import com.google.android.material.tabs.TabLayout;
 /**
  * A placeholder fragment containing a simple view.
  */
-public class TransferRouterFragment extends Fragment {
-    TextView textView;
-
-    private RecyclerView recyclerView;
-    private BedRecyclerAdapter mAdapter;
-
+public class ProfileRouterFragment extends Fragment {
+//    TextView textView;
 
 
     @Override
     public View onCreateView(
             @NonNull LayoutInflater inflater, ViewGroup container,
             Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.transfer_router, container, false);
+        View view = inflater.inflate(R.layout.profile_router, container, false);
+//        textView = view.findViewById(R.id.section_label);
 
-
-        // Set Up Tab
         ViewPager viewPager = view.findViewById(R.id.viewpager);
         setupViewPager(viewPager);
 
@@ -43,16 +37,15 @@ public class TransferRouterFragment extends Fragment {
         TabLayout tabs = view.findViewById(R.id.result_tabs);
         tabs.setupWithViewPager(viewPager);
 
-
         return view;
     }
 
-
+    // Add Fragments to Tabs
     private void setupViewPager(ViewPager viewPager) {
 
         SectionsPagerAdapter adapter = new SectionsPagerAdapter(getActivity(), getChildFragmentManager());
-        adapter.addFragment(new IncomingTransferFragment(), "Incoming");
-        adapter.addFragment(new OutgoingTransferFragment(), "Outgoing");
+        adapter.addFragment(new IncomingTransferFragment(), "My Profile");
+        adapter.addFragment(new OutgoingTransferFragment(), "Change Password");
         viewPager.setAdapter(adapter);
